Organización de los archivos con el modelo MVC
