+++ src/db.js
@@ -1,0 +1,16 @@
+// Conexión a MySQL según tu .env
+
+const mysql = require('mysql2');
+
+const db = mysql.createConnection({
+  host:     process.env.DB_HOST,
+  user:     process.env.DB_USER,
+  password: process.env.DB_PASS,
+  database: process.env.DB_NAME
+});
+db.connect(err => {
+  if (err) throw err;
+  console.log('🟢 Conectado a MySQL');
+});
+
+module.exports = db;
